explicit cast to int

+++ src/i_video.c
@@ -1105,7 +1105,7 @@ static void ResetResolution(int height)
     // 1278x720.
 
     double vertscale = (double)actualheight / (double)unscaled_actualheight;
-    video.width = ceil(video.unscaledw * vertscale);
+    video.width = (int)ceil(video.unscaledw * vertscale);
 
     video.width = (video.width + 1) & ~1;
